Update 1176 - Fibonacci em Vetor.c

+++ 1176 - Fibonacci em Vetor.c
@@ -1,10 +1,11 @@
 #include <stdio.h>
 
-void fibDin(int indice, long int vetFib[]);
+
+void fibDin(int indice, long long int vetFib[]);
 
 int main()
 {
-    long int vetFib[10000]={0,1}, numCaso, indiceDesejado, indiceMaisAlto=1;
+    long long int vetFib[10000]={0,1}, numCaso, indiceDesejado, indiceMaisAlto=1;
     
     for(int i=2;i<10000;i++){
         vetFib[i]=4;
@@ -19,14 +20,11 @@ int main()
             fibDin(indiceMaisAlto+1, vetFib);
             indiceMaisAlto++;
         }
-        printf("Fib(%d) = %d\n", indiceDesejado, vetFib[indiceDesejado]);
+        printf("Fib(%d) = %llu\n", indiceDesejado, vetFib[indiceDesejado]);
     }
-    
     return 0;
 }
 
-void fibDin(int indice, long int vetFib[]){
+void fibDin(int indice, long long int vetFib[]){
     vetFib[indice]=vetFib[indice-1]+vetFib[indice-2];
 }
-
-/* Codigo aprensenta "Compilation error" */
